Fix docs and tests #29
